Separated imageviews with different concerns
