+++ wallet/res/layout/activity_crop_image.xml
@@ -1,13 +1,23 @@
 <?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout
-    xmlns:android="http://schemas.android.com/apk/res/android"
+<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
     android:layout_width="match_parent"
-    android:layout_height="match_parent">
+    android:layout_height="match_parent"
+    xmlns:app="http://schemas.android.com/apk/res-auto">
 
-    <de.schildbach.wallet.ui.dashpay.widget.CircleCropTouchImageView
+    <de.schildbach.wallet.ui.dashpay.widget.BlurredImageView
         android:id="@+id/background"
-        android:scaleType="fitCenter"
         android:layout_width="match_parent"
         android:layout_height="match_parent"/>
+    
+    <de.schildbach.wallet.ui.dashpay.widget.CircleCropTouchImageView
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintTop_toTopOf="parent"
+        android:scaleType="centerCrop"
+        android:background="@android:color/transparent"
+        android:id="@+id/circle_crop"
+        android:layout_width="300dp"
+        android:layout_height="300dp"/>
+
 
 </androidx.constraintlayout.widget.ConstraintLayout>
